[FIX] sap.m.DateTimeField: new method getDisplayedPlaceholder introduced
Until now if the placeholder was not set by the app dev, we internally
visualized as a placeholder the value of the displayFormat. This leads
to confusion whether getPlaceholder have to return it or not.
Now getDisplayedPlaceholder returns the placeholder, no matter if it's
the displayFormat or a set value.

The fix affects sap.m.TimePicker, sap.m.DatePicker & sap.m.DateTimePicker.

Change-Id: I162bbc6fda1460067bb9799e65554af1f0e29b26
BCP: 1880157233