[FIX] core/AppCacheBuster: cache buster token for XHR
With the change to the new loader it occurs that urls which should be rewritten
to have a cache buster token included miss it. With this change each XHR is now
checked for cache buster relevance and the url is rewritten accordingly.

Change-Id: I91dea7dc5b5cea2bd33bb10d78fe880bf26006c8
BCP: 1870010326
(cherry picked from commit cfc51c70cdddebe567ac7dda21a92ab739e6733e)
CR-Id: 002075125800001504432018