[FIX] ui.layout.Grid: Control visibility properties behavior fixed
BCP: 1830101988

Problem: There is some old logic, causing controls to be shown on given
screen sizes, even when their corresponding visible property is set to
false. Also there is a mixture of visible and hidden classes set which 
is confusing.

Solution: Use straight-forward approach to set only a hidden class for 
the given screen size if the corresponding control's visible property 
is set to false

Change-Id: Ie2c68594e16c7241b74183dc82dc6ecdc8c76690