[FIX] MessageManager: Memory leak
When creating a model it gets registered at the 
MessageManager so the MessageManager holds a 
reference to all models. 
If the models are not destroyed correctly we run
into a serious memory leak. 

The registration call is still in place, but the
MessageManager no longer keeps a reference to the
processors (models).

We still have a reference to the Processor (model)
within the Messages. So we still need to rely on 
a destruction of a model (involved in messaging)
to remove those Messages and cleanup the reference. 
Models not involved in messaging are no longer referenced.

Normally models that are involved in messaging
are created by the manifest and so get destroyed 
correctly. In the case an application creates own
models for use with messaging, these models must be
destroyed correctly by the application.

open issues: 
- ControlMessageProcessor has no lifecycle defined,
but stores a map of Messages. 
- need some logging/support tool support to easily
find non destroyed models

Change-Id: I71498c326cf08fc45100dc9e34815bca07857515
BCP: 1880233821
CR-Id: 002075125900000909612018
(cherry picked from commit ad1a86cf2e5fd3c27ccdfd3f529f845568a3f443)