[FIX] sap.m.PlanningCalendar: 1 Month view navigation between months is fixed.
When PlanningCalendar is in "One Month View" and 2 months are displayed,
navigation with arrow right from the end of the first month to the begin
of the second month fails - the dates row is rendered as expected, but
the appointments are not.

The reason for this was wrong and redundant setting of the start date
of the calendar rows to undefined.

Change-Id: Ib4dae238a9eb696f8e537db4208563640722037c
BCP: 1770543942