[FIX] sap.m.PlanningCalendar: customData for PlanningCalendarRow is fully supported.
If custom data with property writeToDom=true is set to PlanningCalendarRow(PCR)
this used not to work, because the data wasn't passed to the internally
created sap.m.ColumnListItem, which corresponds for this PCR.

This behavior is already supported by various container controls, which
have their aggregation item of type sap.ui.core.Element (e.g. sap.m.Table)

Now this is supported.

Change-Id: Ib3493471dd8d2fd7d2d9fe5023898214bc87cbcf
BCP: 1870027607