[FIX]: Opa: option for async polling in success case
A poll interval was skipped after successfull check() and directly the
next waitFor.check() was executed. This did not give chance for app to
process the interaction before the next check.
The issue is seen an interacting with popups while they are not
completely rendered and ready to accept interactions. So interactions
are effectively discarder.
This problem was making some logically wrong scenarious to succeed.
Such scenarious would need validation and fixing.

Change-Id: Ie824a225bee657bee601f158a6b165ea1abdb278