[INTERNAL][FIX] Configuration: Provide CSP compliant alternative for config parsing
To allow for a CSP compliant config parsing we now first make a
JSON.parse try.

If this fails we still fallback to the not compliant
"new Function" eval.
This might happen in legacy cases when code should be executed
during config parsing or when the config is not given in valid JSON syntax.

Change-Id: I081c608d2f552bdd1f5d5d150978d1e0689b159f
JIRA: CPOUIFPHOENIXCORE-2461