[INTERNAL] sap.m.Select: Internal input introduced
Since sap.m.Select is rendered as div in the DOM, rather than
a select, this causes sap.m.Label to render as span, rather
than label. This was done in order to avoid invalid HTML being
output by our controls. By doing so however, upon clicking on that
span, focus won't move to the sap.m.Select (which is standart
label functionality).

To fix that, we:
 - Render a hidden input inside the sap.m.Select
 - Return the input's ID in getIdForLabel() method
 - Remove sap.m.Select from the non-labelable controls in
   LabelEnablement, since the label would now reference an
   input element, which means HTML validity is achieved.

In addition to this, we:
 - Indicate whether the control is enabled or disabled
   via aria-disabled

Change-Id: Ie4aac01c183a5efe8a7681ed4c9eb898e3bfae71
JIRA: BGSOFUIPIRIN-1892