[FIX] sap.m.TileContainer: active page is displayed correct now.
The active page in TileContainer (TC) is usually marked as a color filled
dot. When user is on the second page, then switch TC visibility off,
then on, the pager shows that both first and second page are active.

The issue was wrong expectation about the index of the second page.

Change-Id: If2098037acee891386e912a10965c912915334a4
BCP: 1880199566