[FIX] sap.m.TileContainer: after new page is added, scrollers remain visible.
If the end-user is at second page, which is also last, and then some tiles
are added so there is another page created, switching off the visibility
of the TileContainer and then restoring it back to on, leads to disapeareance
of the navigation buttons.

The fix relies on the following logic:
Whenever the pager is created from scratch, update of the scrollers is needed.

(manually cherry picked from c3296c685651c79c44629b3cbc605ed6d4596be5)

Change-Id: Iad5d81e5b0f30516e02889bbef1768dee65b44a2
BCP: 1870041307
CR-Id: 002075125800001677502018