[FIX] sap.ui.fl: Variant selection in URL should not change history
When selecting different variants in an app containing variant
management controls, the URL is changed correspondingly. These changes
should not be recorded in the browser history, so that pressing the
"back" button will return to a previous page and not simply go back to
the previously selected variant.

BCP: 1870040392
CR-Id: 002075125800001751022018

Change-Id: I8af6c8ee32f7be19f268099e094756919c00367c
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>
(cherry picked from commit a1632670edb941c451dfb819e0dfcb1975008215)