[INTERNAL] ControllerExtensionMetadata: First draft
new ControllerExtensionMetadata to control method
visibility/final state.

All methods of an extension prefixed with '_' are 
private. All other methods are public. The visibility
can be overridden by the new metadata section:

metadata: {
	methods: {
		"myMethod": {public: true, final: true}
	}
}

Methods flagged final cannot be overidden by another
extension. An error is logged on the console.

- Member extensions are directly extended in the 
constructor of a controller
- the getPublicMethod function of the controller is
enhanced with the new metadata
- the base member of an extension now also contains
all the exetensions
- An Extension class provides a static override function
	- introduced as we don't want allow extending 
	ControllerExtension 
- Extensions are flagged as final and can't be extended
- Extension has a static override function:
	- as Extensions cannot be extended use
	the override function instead:
		DraftExtension.override({
			functionToOverride: function() {
			...
			}
		}

Change-Id: Idde936ca908f4386b17ac042026fb31fc6615891