[FIX] jquery-mobile-custom: Edge button events
This change implements a workaround for an Edge browser bug which occurs with
EdgeHTML15 and higher. When tapping a button, the event targets are differ,
if another html tag is embedded in the button tag. This leads to not fired
tap events on the sap.m.Button control.

BCP: 002075129500005680192017
BCP: 002075129500001076112018
BCP: 1880219029
BCP: 002075129400004318102017

Change-Id: I78dae654aad4036774b1e3f4ab07b4d5d84423f1