[FIX] Testing: use newest release of sinon to fix sinon issue #1534
The sinon release 4.1.3 - which is currently packaged with UI5 as
sinon-4 - has an issue when a fake server is created from a sandbox. The
created server unintentionally fakes timers as well and therefore
doesn't auto-respond.

See https://github.com/sinonjs/sinon/issues/1534 for details.

Sinon release 4.1.6 fixes the issue, this change applies the newest
available release 4.4.6. 

Change-Id: If9207574f1e0b7d253ecb15041b15363086fa541