[FIX] mvc/XMLViewRenderer: create DOM string before dummy placeholders
It occured that Control's BeforeRendering delegates have not been called
properly, if the XMLView replaced the Control's DOM with placeholders
before RenderManager.renderControl for certain children was called. That
led to a state, where in the BeforeRendering delegate a Control, which
actually was rendered earlier, could not find its DOM reference via
Control#getDomRef.

Change-Id: Ib8183265f5eb2fbd5c46da8a10467c519fdce0ee
BCP: 1870027560