[FIX] ComponentContainer: Convert "manifest" string to boolean
In case of using the ComponentContainer within an XMLView the conversion
of the string values into a boolean is not working as the type of the
manifest property is "any". Therefore the type conversion will be done
in "applySettings" which handles properties provided initially.
The setter (setManifest) will still work as before and won't convert a
string into boolean, as this isn't required for the XMLView use-case.

Change-Id: I4ea5d6f3751e22b4c13dff33f4b2aa6e975acc82
(cherry picked from commit f626a8007cffc664460aebe218789890b1985b94)
CR-Id: 002075125900001014792018
BCP: n/a