[FIX] sap.m.DateTimePicker: when picker is opened for IE & Edge, input should not be focusable
For IE & Edge, any selection of the underlying input must be removed before
opening the picker popup, otherwise the input will receive focus via TAB
during the picker is opened. The selection is restored back when the popup
is closed.
This is also aplicable for DateRangeSelection and DateTimePicker
This change is continue of the Ia8c26aadeedc6690094e63ecd1c5d95b8f0c8670

Change-Id: Ia194d0927060559f5d0c2fa5c590e0cbaaee0163
BCP: 1880193676
CR-Id: 002075125900000857562018