[FIX] uxap.ObjectPageLayout: footer content alingment fixed
Issue
The vertical alignment of the sap.m.Bar content is wrong,
when used as floating footer

Root cause
The sap.m.Bar has height 2rem in compact (by spec),
but the OPL makes it 2.5rem (also by spec).
The Bar content alignment depends on having height of 2rem,
this changing it to 2.5rem breaks the vertical alignment

Solution
Set the desired 2rem of height to the FloatingFooter,
which will fixes all alignments in the Bar as this is the height in the Bar`s spec.

Note: The price is to have wrong visual (height 2rem istead of 2.5rem)
of the Floating footer in compact mode with sap.m.Bar.
If this combination turns out to be widely used we should go for changing the Bar`s styling
in the context of Floating footer.

BCP: 1870106645
Change-Id: Ib8676ae590976b6a98aec3d32938c22673c113b1