[FEATURE] sap.ui.Device: Introduce Blink flag
The Chrome browser from version 28 and above and various other browser use the
Blink rendering engine. But often the user agents still includes the Webkit
keyword.

For compatibility reasons the Webkit detection works like before, but a
new blink Device API flag is introduced, which is set independently of
the Webkit flag.

BCP: 1870097430

Change-Id: I99555b0ee7b6e12e0514b136b9c31bc03862e92d