[FIX] jQuery.sap.includeStylesheet: Correct comparison with existing href
The comparison of the old and new href didn't correctly work in all
cases. This caused reloading or invalidation of stylesheets as the
actual URL (or href) was the same but the <link> tag was still replaced.
When accessing the "href" of a <link> tag, the browser already resolves
the given href as absolute URL which can be compared against the
existing href.
This also makes the use of the URI.js library obsolete.

BCP: N/A
CR-Id: 002075125900001060752018
Change-Id: Iddc60409b09221e71afb696b4e1f174c925b3c6c