[FIX] XMLComposite: Cloning of XMLComposite was not possible
Note:
As _content was private it was not tried to be cloned, this is changed
in the metadata.
Now as that can be cloned it suffices to set this managed object as
content. Tested that already in the feature branch for new mdc field.

Change-Id: I2acddbd5937377f75dfbbacf1d76fb1d556e344c
(cherry picked from commit 29fc8e65929c43448f841ae8e5601addc050f606)