[FIX] sap.f.DynamicPageTitle: aria-labelledby now properly updated
The aria-labelledby attribute was put on the DOM element regardless
of the parent's toggleHeaderOnTitleClick. Now this attribute takes
into consideration toggleHeaderOnTitleClick's value and also
updates the references when it is changed.

In addition to this, several calls to the resource bundle for text
"COLLAPSED_HEADER" are now properly changed to search for "SNAPPED
HEADER"

Change-Id: I42e823235aa51dbf4650177300049fb28427384b
BCP: 1870102287