[INTERNAL] Event handler parameter syntax: controller func without dot
Let the expression parser apply the same logic as the handler resolver
always used: even with no leading dot: when there is no other dot and
the name represents a member of the controller, use this mamber as
handler.

Also remove some trim() operations which would make the overall behavior
more tolerant, but is not consistent with the rest of the
implementation.

Change-Id: I264b612bfa64702d641216a1d6470f4f64d15a72