[INTERNAL] ui5loader: deactivate async error handling tests in FF
The support for document.currentScript differs between Chrome and Safari
on the one hand side and IE and FF on the other side. IE does not
support currentScript and in FF, the currentScript during a
window.onerror listener is no longer the script in which the error
occurred.

This makes it impossible for the ui5loader to assign an uncaught error
to the module in which it occurred. 

Therefore, the corresponding tests are now deactivated for Firefox as
well (they had already been deactivated for IE).

Additionally, the code that loads a failing module (loaded via XHR) a
2nd time (via script tag) to facilitate error analysis, has been moved
to the correct place.

Change-Id: Iad570e99faa20b6bc1fdb3f3bb015360ec52d3fb
BCP: 1870141297