[INTERNAL] sap.m.Table: Focus border visibility and focus handling fix
A new class is now added to the tbody element of the table which has
the white background color and the table list items are made
transparent. This is done to ensure focus borders are visible on the
header row of the table.
The top border of the first table item is also removed as it was
overlapping the focus border of the header row. The issue was only found
with the belize theme and hence the fix is only done for that theme.
The focus handling was adapted to work with the heights of the sticky
elements instead of last sticky element which ensures reliability.
Resolves:
JIRA: CPOUIFTEAMB-1121

Change-Id: Ic9194ec47902177b1fbc4b4f207688810066f271