[FIX] sap.m.PlanningCalendar: correct display of the Week with maxDate
When maxDate is set to the PlanningCalendar in Week view it tries to
correct the maxDate in order to display the week from the first day of the week.
So it substracs from the maxDate the number of days to display + 1 and this becomes
the max date.

Now we correct this functionality and in Week view we do not correct the max date so
it always can show the week from its first day of the week. The days that are outside
of the maxDate will be shown and will have no action when click them.

Change-Id: I5c7c81e20579778d0f10f6e4f178cf7a4e2d6a30
BCP: 1870161990