[FIX] sap.f.DynamicPage: footer animation corrected
Background
Under specific circumstances of race condition of
showing and hiding the footer leads to wrong footer appearance.

Root cause
The showing of the footer is executed immediately,
while the hiding is performed in a timeout.

Due to binding resolution, the flow of the program might be:
toggleFooter(false)- seting a timeout to hide the footer
toggleFooter(true)- immediately showing the footer
The timeout is executed later and hides the footer, althought it should be visible.

Solution
Clearing the timeout every time the footer is about to be displayed.

Change-Id: Ib4115e035f290c714d481f94814fee487ae8f029
BCP: 1880089543