[INTERNAL] AggregationProxy: delete
-we have deleted the AggregationProxy because it has
been replaced with the more generic approach of
'forwarding' in the ManagedObject. Since there is
still a discussion pending regarding the model
propagation when forwarding is defined, we have
deleted an example for now. This example can be
added again (modefied form) if we decide to propagate
models of elements which are forwarded.

Change-Id: Icab303d1bbfb7eeb6ebe55ccf3be4d336c784142