[FIX] sap.ui.rta Reload page for app descriptor changes
App descriptor changes are processed on the server, to get an
immidiate feedback there are runtime only changes applied. But with an
FLP cross app navigation, which can already occur to get personalization
changes, the runtime only changes are no longer there. A real page
reload is needed for a new app descriptor to take effect.

Change-Id: I8578da191b65d8ff827c94ac4873415c9a63fdff