[FIX] sap.m.Image: DetailBox opens when Image is used with list binding
The problem:
When the Image control is used in a list for example in a
CustomListItem utilizing list binding and a sap.m.LightBox control is
added in the detailBox aggregation the press event does not open the
LightBox control.

The root cause:
List binding internally clones the template control and the LightBox in
the aggregation but the "press" event listener points to the instance
used by the list binding as original/template
(which is not rendered in the dom) thus clicking one of the rendered
instances of the Image calls the listener on the one used for template
is called and this fails with a side effect error of not having
imageSource.

The solution:
Override the clone method to transfer the correct event listener to the
cloned instance of the Avatar control.

CR-Id: 002075125900001534512018

Change-Id: Id95f3885da4b92405a13605b4380d558b99c00b5
BCP: 1880004603
FIXES: https://github.com/SAP/openui5/issues/1825
(cherry picked from commit 4ac1f7ed0bf210e32f4dd378346058635d9219c2)