[INTERNAL] sap/ui/model/odata/v2/ODataModel: Remove generated uri from request
Remove generated uri from createEntry request.
The uri which is sent with every request upon
ODataModel.prototype.createEntry call is generated.
Therefore it doesn't point to a valid resource.
This causes an issue for some backends which validate the uri sent with
every request. Therefore the uri is removed before sending.

BCP: 1870027369
Change-Id: I171f8cac66c6cc679b3a8f95479d5b85f7269cdf