[FIX] v2.ODataListBinding: Implicit casting for filter values
Due to the usage of OData comparator methods, camparison doesn't do
implicit casting of filter values anymore. This is on the one hand
imcompatible to previous behavior, on the other hand differs from the
behavior of server side filtering.

Removed "getFilterComparator" which solved this issue for Edm.Decimal
only and replaced it with "normalizeFilterValues", which does type
casting as necessary for all number types.

BCP: 1870002694
Change-Id: I24af09645be50875a732389701a2018ea3766d9a
CR-Id: 002075125900001694762018