[INTERNAL] sap.ui.fl allow changes to be applied on js only
Some controls API don't match between JS and XML. In these cases e.g.
indexes will not match. With this change RTA can detect these cases
via designtime metadata and stores that information in the change.

As the controls that have this problem at the moment need to propagate
this information to other controls we want to have the flag in
designtime.

Change-Id: I535f2f9f9154c87b10d582d8611ceae721cd9362