[INTERNAL] sap.ui.core.Popup: adding delay to popup blocker when sapUiBLyBack style class is removed from html element
When the sap.m.Table is configured with sticky header elements, the
removal of the sapUiBLyBack style class lead to scrolling of the page to
the top and due to this the sticky position of the header elements was
not able to be retained. Setting a delay helps the sticky elements
retain its position after the DOM is done repainting the page. The
scrolling issue was observed in Chrome only.

Change-Id: Idfa5a983618b1f052ce28444911f81d25d3d29e6