[INTERNAL][FIX] sap.m.Slider: Take into account Tooltips' formatter for accessibility.
In order to format values we relied on Slider values or Scale's formatter.
However, there might be cases where only custom tooltips are used and
Screen Readers were not able to provide proper output.
That's why we need to take into account also the tooltip formatters (if provided).
With that change there's a priority in the used formatters as
the default behaviour is as follows:
- Slider value
- Scale formatter (if any)
- Tooltip formatter (if any)

That behaviour could be justified and controlled with a parametrized prioritization.

JIRA: BGSOFUIRILA-2102

Change-Id: I66ce4aa8fab1a30a5c9994ff8045dfdf400d6cbb