[FIX] sap.m.TileContainer: reposition tiles when resize happens
Previously reposition of the tiles occurs only if maxTiles is different
before and after the resizing. However, when a small resize of the browser
window (or zoom in/zoom out) happens in the browser it is possible
again the old and new maxTiles to be the same number. This will cause
some visual issue like the tiles which will not be centered on the screen
until it is clicked and reposition logic is forced.

Now in resize handler if the size of the TileContainer is changed the logic
for repositioning the tiles is applied

Change-Id: Ifbdb2afcadb280c0672b960a14f3e75124dbe791
BCP: 002075129500001515232018
CR-Id: 002075125900001868782018