[FIX] sap.ui.documentation: Fields table visibility improved
The problem:
In the samples section of the DemoKit the Fields table was visible in
the about section of controls having fields for example "sap.m.Text".
In the API Reference page the Fields table did not show the visibility
info about the members

The root cause:
The table in the sample section about tab was a left over from the old
explored app. The visibility of the members in the API Reference section
was not implemented

The solution:
Remove the table from the samples about section and added the visibility
status indicator in the API Reference section

Change-Id: I04cd0921891688cf30a671ce496d8fae699147c4
BCP: 1870230751