[FIX]sap.ui.rta - Add elements dialog model contains circular references
In the Additional Elements Dialog we keep a list of entries in a JSON
model. In the case of reveal, an entry includes a reference to the
element itself. This creates circular references in the JSON model that
in certain circumstances cause an exception when executing stringify.

Since we don't need this reference in the dialog, this change replaces
it with the id of the element.

CR-Id: 002075125900001961102018
BCP: 1870253528

Change-Id: I900fb0405655b94149efd079a293b1f410c38e68
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>
(cherry picked from commit f8b94edd3d4dab0a1b2e2fca4125ebb600d3506f)