[FIX]sap.ui.rta - Add elements dialog model contains circular references
In the Additional Elements Dialog we keep a list of entries in a JSON
model. In the case of reveal, an entry includes a reference to the
element itself. This creates circular references in the JSON model that
in certain circumstances cause an exception when executing stringify.

Since we don't need this reference in the dialog, this change replaces
it with the id of the element.

Change-Id: I900fb0405655b94149efd079a293b1f410c38e68
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>