[FIX] sap.m.MaskInput: on Android, caret positioning is improved.
On Android devices with Chrome, if the user enters a value in the middle
of the input, the value is being recognized at the beginning of the string.

Fix hints:

The root issues is that there is no a keypress event for this combination.
(https://bugs.chromium.org/p/chromium/issues/detail?id=118639).

To workaround it, we simulate the keypress by attaching on the appropriate
places in order to fully mimic the flow that the other devices have.

Full description on the issue and the fix is attached as js doc on the
method _onInputForAndroidHandler.

Fixes: https://github.com/SAP/openui5/issues/1843
BCP: 1880062950

Change-Id: Ib6d8733717a8b926a53d7229f44869ad053eed2d
(cherry picked from commit 975844ca461e393330f21a7f271e0f40e31c09f4)

Change-Id: I16fb271794e7088f82f1ea11914444ac7c43827b
CR-Id: 002075125900002210052018