[FIX] CompositeBinding: Use validation of nested bindings
When a CompositeBinding is using raw values, but has constraints defined
in types of nested bindings, these constraints also need to be checked.

This is especially needed, when Currency is using in combination with
Edm.Decimal, where minimum/maximum constraints on the Currency itself
cannot be used because they don't have the necessary precision for
Edm.Decimal.

BCP: 1870235627
Change-Id: I7a66210d1eff8053c397e7955c60b03303ef4e0d