[FIX] ui5loader: don't modify global define/require props by default
To avoid conflicts with UI5 consumers that already use another AMD
loader, the UI5 loader by default no longer acts as an AMD loader. That
is, the global properties window.define and window.require are not
modified. 

Only when the new config option 'amd' is set to true, the properties
will be set to the corresponding implementations of the UI5 loader.
Switching the option back to 'false' will restore the previous
implementations.

The config option 'noConflict' (named 'noLoaderConflict' as URL
parameter or script attribute) is still supported, but deprecated.

Change-Id: Idce249414802c1850f9870679448031f06984aba
BCP: 1870246244