[INTERNAL] sap.ui.fl: copy dependencies when control gets recreated
until now FL expect a control to not be destroyed and recreated as
long as the component is loaded, because we only create the dependencies
once on component startup. Then when creating the control for a second
time the dependencies are missing and changes are applied without.
With this change the dependencies are getting created when recreating
controls

Change-Id: I5ae82c792da0792654bf9fc85debe8df09344ea5
Signed-off-by: Kevin Edinger <kevin.edinger@sap.com>
BCP: 002075129400002261702018