[FIX] sap.ui.documentation: DemoKit robustness improved
The problem:
When certain symbols are opened in DemoKit it fails to open their page
and freezes on the loading phase with no meaningful error in the console

The root cause:
On two places in the xml template the method name is used as ID for UI5
control. Some methods are returned with "#" character in the method name
which is not permitted as ID character. Also exception handling needed
improvement to show more meaningful error messages

The solution:
Removing possible "#" characters from method name when used for ID in
the xml template. Improved error message reporting

Change-Id: I2a93bc8a3f2146bf05d3aadfec16ec212f94ada3
BCP: 0020751295 0000243591 2018