[INTERNAL][FIX] sap.m.ColorPalette - less parameters names fix
The names of the parameter should follow the guideline:

For control-specific parameters in *.less files, use a combination of
the library name and the *.less file name for the prefix.
Start with an underscore. Separate each part of the library namespace
and the file name from each other using underscores as well.

Local parameters themselves must not contain underscores.

Change-Id: Icbbf6589c3577e73fc2effe7d6e16eaee3b1dafe