[FIX] sap.uxap.ObjectPageLayout: ScrollEnablement now has latest DOMref
The ScrollEnablement obraines a reference to the latest domref of
ObjectPage in a delegate to "onAfterRendering" of the ObjectPage

However, the delegate is called after the "onAfterRendering" hook of
ObjectPage itself => any scrolling called there fails
as ScrollEnablement has old reference

ScrollEnablement uses fallback (caches latest request change
in scroll position, in order to restore correct position once
it obtains the latest domref).

However, the fallback fails on Mozilla due to race condition.
This change demonstrated a workaround in case the Core colleagues
refuse to fix in the ScrollEnablment itself (deprecated already).

Change-Id: I3206b47b5a0234c0dcd18614140fe2882664b8e9
BCP: 1870201875