[FIX]sap.ui.dt - Removed overlays are still part of the selection
After an overlay is destroyed (e.g. one of the combined fields after the
"combine" action), it is still part of the selection until the selection
manager "remove" function is called. This was not handled properly and
the selection was temporarily incorrect, corrupting the actions
available on the context menu for the selected existing overlay.

This fix removes the destroyed overlay from the selection, allowing the
remaining overlay to stay selected after "combine" and with the correct
actions enabled on it.

BCP: 1870313033
CR-Id: 002075125800001973462018

Change-Id: I97bc530828076b56aad46f9e9d35c1d0f42384f3
Signed-off-by: Vitor Seifert <vitor.eduardo.seifert.bazzo@sap.com>
(cherry picked from commit f8ba865568e51cf165243d260c217bb369d1b5bf)