[INTERNAL] sap.ui.fl: copy dependencies when control gets recreated
until now FL expect a control to not be destroyed and recreated as
long as the component is loaded, because we only create the dependencies
once on component startup. Then when creating the control for a second
time the dependencies are missing and changes are applied without.
With this change the dependencies are getting created when recreating
controls

Change-Id: I2d8776379fef21b8008557e10b6b7a4f5a5214b6
Signed-off-by: Kevin Edinger <kevin.edinger@sap.com>
BCP: 002075129400002261702018
(cherry picked from commit 37cd469c1853537e289a95959b6301ab042d938f)
CR-Id: 002075125900002328212018