[FIX] sap.m.FacetFilterList: Calls to the backend are now optimized
This reverts commit 0ff7ffb5a7d0d67ac5de20b565c7e48030c05e07.

The reported incidents were because of application code. That's why we revert
the provided fix and add a fix for an issue regarding the optimization of
the backend calls. There were issues in the search function when searching
for an empty value.

Change-Id: I1e221a5e10d91ef5c5cf86536760263336732d77
BCP: 002075129400002300782018
(cherry picked from commit ae0049eff7bfb99af891953f059cc1bec59b0447)
CR-Id: 002075125900002414352018