[INTERNAL] sap.ui.fl: copy dependencies when control gets recreated
until now FL expect a control to not be destroyed and recreated as
long as the component is loaded, because we only create the dependencies
once on component startup. Then when creating the control for a second
time the dependencies are missing and changes are applied without.
With this change the dependencies are getting created when recreating
controls

Change-Id: I91a0c2b194c791d623b94ff18ed2215e233508a8
Signed-off-by: Kevin Edinger <kevin.edinger@sap.com>
BCP: 002075129400002261702018
CR-Id: 002075125900002328222018