[INTERNAL] sap.m.Table: sticky scroll handling
When the display settings of the screen are greater than 100% scale,
this resulted in unexpect scrolling of the page when stickiness is
enabled in the table. To avoid this, the float values returned in the
_handleStickyItemFocus() are converted to integer value to correctly
compare if scrolling of the page is necessary.
BCP: 1870215318

Change-Id: Ic3efdd73305606bde774070f3dbfd974268358f9
(cherry picked from commit 0fdba60544fc476f4bb883ff2d143c5cfe672b0c)
CR-Id: 002075125800001980522018