[INTERNAL][FIX] Icon: Fix legacy 'src' without Icon URI
Although it is not mentioned in the API, the 'src' property also
accepted a icon name of the default collection as value instead of the
documented Icon URI (IconPool.getIconURI).
This is was caused by 867c4c69984835d514d5a507c8f8701f915fa8a6 and works
again, but is still not recommended. Therefore an assertion is added.

BCP: 1870325627
Change-Id: I9d88d26f2b23cc021675238b1efb8aa0d6c0d088