[FIX] GridTable: NoData handling for $expand(ed) lists corrected
If data in the table was being filled as a result of $expand and not
directly due to bindRows on the Table, the content was updated
correctly, however noData text was not removed.
In $expand scenarios there is no event from the binding when data is
requested/received, hence we shouldn't merely rely on those events.
NoData text should basically be updated whenever table data/rows is
updated and there is no pending request, regardless of the kind of
binding (e.g. client).

BCP: 1870263895
Change-Id: I68d3b5c12f3183dc602dd35ffa11ed7a47ae1a44