[FEATURE] DynamicPage: backgroundDesign property added
Add new backgroundDesign property to DynamicPageHeader and
DynamicPageTitle controls, which determines the background color.
In order the new property to be supported correctly in ObjectPage,
also add new backgroundDesignAnchorBar property to ObjectPageLayout,
which is used to control AnchorBar's background color.
Create an observer for HeaderTitle changes of backgroundDesign in ObjectPage,
which is used to update at the same time the background color of the
HeaderContent.
Update units tests. Create sample for Demo Kit.

Change-Id: I0ee07172f7500fee6dbbd34182122ab97d2c3f84
JIRA: BGSOFUIPIRIN-2705