[FIX] XMLComposite: Allow templating inside the composite
Note: Allow a templating inside the composite like template repeat in
columns. The prioer change here yelled the problem, that _content is no
public aggregation and the visiting dumped without a message. Now we
first visit the child nodes, template additional aggregations and then
add the content. This still works with the Table0.

Change-Id: I356c16e26f91ae60a8592bbe44cae19ea525ef0a