[INTERNAL] odata.v4.OData(Context|List)Binding.qunit.js: avoid unlock tasks
In their constructor these bindings create a read group lock and add a
prerendering task to unlock it again in case it is not used. This unlock
task however is not needed in the tests and may interfere with other
tests, especially in IE11. So we prevent that the unlock task is
created.

PS1: done
PS2-3: review comments
PS4: retrigger voter

Change-Id: I9d9fecf3d27c7cb345ec609b33b459fc386a9744
JIRA: CPOUI5UISERVICESV3-1332