[FIX] Demo Kit: Updated filter search "not found" message
Previously, the filter search "not found" message was "No data".

Now, it's "No data, please try with the global search".

Additionally the filter icon in the "Samples" section is changed and
the layout of the search field is fixed, so it's identical to the
adjacent "API Reference" and "Documentation" tabs.

Change-Id: Ia5640f8545fd41f8d268ae34b888ef3ceef2d595
JIRA: BGSOFUIPIRIN-2837