[INTERNAL] flexibility diagnosis panel enhancement and app
This commit refactors the "Flexibility" Diagnosis panel.
Concrete is the table with the applied changes gone.
The lists for all loaded (listed) applications are now downloadable.
A link exists pointing to the corresponding display application.

PS 2: renaming of the app & removal of json cycles in the captured data
;)

PS 3: Refactoring to unit test the app.

PS 4: final tests added.

Change-Id: I2c5e42cb3c017429a96c29d538a722c02d0d37c6