[FIX] sap.ui.model.analytics.AnalyticalBinding#_createMultiUnitRepresentativeEntry
In multi unit situations a new key is calculated for those entries.
The key is a comma seperated list of all loaded dimension values. The
key is stored at __metadata.uri. If the dimension values contain a "/"
it acts as a seperator in __metadata.uri and the keys in the model may
not be unique anymore. Therefore all dimension value are URI encoded
before they become part of the new key

PS1: downport of #3419093

Change-Id: Ie1709459a319028bb700a117b6eb89c7d756782b
BCP: 1880292532
CR-Id: 002075125800003198892019