[FIX] sap.ui.model.analytics.AnalyticalBinding#_createMultiUnitRepresentativeEntry
In multi unit situations a new key is calculated for those entries.
The key is a comma seperated list of all loaded dimension values. The
key is stored at __metadata.uri. If the dimension values contain a "/"
it acts as a seperator in __metadata.uri and the keys in the model may
not be unique anymore. Therefore all dimension value are URI encoded
before they become part of the new key

PS1: initial
PS2: comment SBE
PS3: rebased
PS4: comment TCH

Change-Id: Ia34dd65fff62873d5c23c5be250af6567ad1f678
BCP: 1880292532