[FIX] sap.f.DynamicPage: Initially snapped header behavior fixed
BCP: 1880249493

Problem: When the header is snapped initially and the page content is
lazy loaded afterwards, we get in a case, that a hidden class is not
deleted from the header upon epxanding on user interaction.

Solution: Make sure the hidden class is removed in this case too. qUnit
added.


Change-Id: I0d6f8ad70524b271e7668c15fa02aa6e00d917a6