[FEATURE] OverflowToolbar: shrinking of controls is enabled
Check for controls with LayoutData with shrinkable = true and minWidth,
and add CSS class for shrinking, if needed.
Keep old logic for shrinking controls without LayoutData, only if they
cannot overflow and there is no place for them in the Toolbar, but they
meet at least one of the other shrinking conditions - width in %, text nodes or
controls, which implement IShrinkable interface.
Update unit tests, test page and samples in DemoKit.

Change-Id: I19686632e444004e80ebd9a251b3f688a20fab9d
JIRA: BGSOFUIPIRIN-2925