[FIX] LocaleData: Return Islamic as preferred calendar for ar_SA
In CLDR for locale ar_SA the perferred calendar is "islamic-umalqura".
In enum CalendarType we only have "Islamic", so it doesn't match and
falls back to "Gregorian". As we don't have support for subtypes of
Islamic calendar yet, this change ignores the subtype and choses
"Islamic" as preferred calendar for Saudi Arabia.

Change-Id: Ibaa847f440a689d68bdb1a539a576989c621fda1