[INTERNAL] XMLComposite: Nesting composites
Note:
The change contains several small adaptations on subartifact which are 
not in direct association to what we test.
The main thing we test here is that the initialTemplating of the cell 
is triggered.
Therefore:
- we needed to garantee that the attributes of the cell 
are visited (registerPlugin in unit test)
- the cells become stable ids that we can access them in a 
unit test(new property key for column in the existing calumns frgament)

Change-Id: I15794d959704db408b7ec5ef8e8d4027a368df22