[FIX] Table: Avoid infinite re-rendering if fixed columns are available
Fixed columns with a variable width are initially rendered with 160px.
If the fixed columns do not fit into the available space, column
fixation is disabled. Because the columns are now rendered with variable
width, they fit into the available space. Column fixation is enabled
again. This results in an infinite re-rendering cycle.

To avoid this problem, the calculations use the width of a column as if
it were fixed.

Change-Id: I9608a7049d6bdbf4e16083ad1c18322abe325a9f
BCP: 1880301265
CR-Id: 002075125800002563732018