[INTERNAL] sap.ui.support: Support Assistant new history formatter added.
There was a problem with ABAP JSON parser which cannot
parse current history version because the objects have to be
arrays.

We have implemented a special formatter which will generate
the wanted output. This formatter will be called within the
public API getFormattedAnalysisHistory function.

For more information see the JIRA item

JIRA: BGSOFUIRODOPI-1262

Change-Id: I2d7112eacba9de264348917008142710e76fadb8