[FEATURE] sap.m.ColorPalette: up, down, home and end keys improvement.
Color swatch items should be navigated in a column like mode.
When first (or last) item is active and [Up] (or [Down]) key  is pressed
the focus moves to the "Default Color" (or "More Colors...") button.

[Home] moves the focus to the first color in the row, or to the first
color in the container or to the "Default color" button.

[End] moves the focus to the last color in the row, or to the last color
in the container or to the "More Colors..." button.

Change-Id: I408d07b9dffb0d5f8f1e53fd143fa413285444b6
JIRA: BGSOFUIBALKAN-4364