[FIX] NumberFormat: Fix currency short format for RTL languages
There are display order issues with RTL languages, if the scale is in
arabic letters and the unit in latin letters. The last try to fix this
rather made it worse, so this change inserts individual control
characters for "ar", "fa" and "he", to make them appear as expected.

BCP: 1880305750
Change-Id: I6b4cc11fc9c4c85d6df89bf2cc4af4e4866027a3
CR-Id: 002075125800002893822018