[FIX] Component: Make url-objects working again in asyncHints.libs
When declaring libs in asyncHints, it was possible to declare URLs as
final, by setting them to a JS object with a "url" property and a
"final" property. As the new loader does no longer support this feature,
this change removes all these objects and replaces them with the
contained URL. This only happens in the old factory function
"sap.ui.component" for compatibility with existing applications.

BCP: 1880339873
Change-Id: I8945505c2538d5b80c8c40c731a583e495ac76f5