[INTERNAL] Form, SimpleForm: use title ID of parent container
The problem:
When we have Dialog -> Form, SimpleForm or SmartForm relationship
and the latter control is the first control in the Dialog and has
no title the screen reader reads every item in the Form multiple times

The root cause:
This is a known issue on certain screen reader and a Form without a
title. Introducing a second title to the form does not make sense in
most cases and invisible text will force applications to modify their
view's source and leaves the risk of obsolete controls in the future

The solution:
We propagate the Dialog title ID to the first child control if it is
of the listed types and implements the suggest method thus enabling it
to mark itself as described by the title of the parent control

The same happens in ObjectPage, Wizzard and other containers too.

Change-Id: I9053cab02cc1e918e73f35ab9e528d3bff07fdfe