[FIX] f.FlexibleColumnLayout: suspending/resuming of ResizeHander fixed
Issue
The content inside FCL columns could not respond to resize of the browser
due to ResizeHandler suspention.

Root cause
The ResizeHandler is resuming only if transition occurs.
But imagine a simple resizing of the browser window, which does not lead to
column show/hide animations - all these interactions will not trigger the
resize handler as it is suspended and expect column layout change to resume.

Solution
schedule a timeout with the sam delay as the animation duration
to resume the ResizeHandler.

BCP: 1880331828
CR-Id: 002075125800002806172018
Change-Id: I09c2ce086f434b7a23f94ce891340955e67ed57f
(cherry picked from commit 8b9b1041021589dddcbbe9169298d7342464d851)