[FEATURE] ui5loader: Add AMD compliant require.config()
The ui5loader can be configured with an AMD-like config
function (sap.ui.loader.config) which does not completely
respect the AMD specification w.r.t "paths" and "baseUrl".
This can't be changed due to compatibility reasons.

When the ui5loader is switched to AMD mode by calling:
  sap.ui.loader.config({ amd: true });

there is now an additional "config" function available as:
  require.config({ ... })

which behaves as described in the AMD Spec:
https://github.com/amdjs/amdjs-api

JIRA: CPOUIFPHOENIXCORE-2542
Change-Id: I5683db21784fb7e267957060ce2ce94c9cc81ecc