[FIX] Component: asyncHints with "final" flag
This is a follow-up of dcc0e7f468ff3eec8ef3607620d508895a1d09bd

With the refactoring in 609974b78136a287ab376b03299653c9459526a0
the Component now uses the loader API to register module paths.
For the "final" flag, the old API still needs to be used as the loader
doesn't support it.

BCP: 1880339873
Change-Id: I333ba24684d1ce0434ffa08eff9621573417d5f2